# src/main/resources/openapi/overview-annual-comparison.yaml
openapi: 3.0.3
info:
  title: Grain Trade Overview Annual Comparison API
  version: v1
paths:
  /api/v1/overview/annual-comparison-definitions:
    get:
      summary: List database-owned annual comparison indicators applicable to a business domain and product
      parameters:
        - {name: sourceDomain, in: query, required: true, schema: {type: string, enum: [PRODUCTION, MARKET]}}
        - {name: productCode, in: query, required: true, schema: {type: string}}
      responses:
        '200':
          description: Applicable approved-fact indicators ordered by enterprise metadata.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AnnualComparisonDefinition'
  /api/v1/overview/annual-comparisons:
    get:
      summary: Read four comparable annual approved-fact values
      parameters:
        - {name: productCode, in: query, required: true, schema: {type: string}}
        - {name: cultivarCode, in: query, required: false, schema: {type: string}, description: Production-only cultivar filter.}
        - {name: regionCode, in: query, required: true, schema: {type: string}}
        - {name: periodCode, in: query, required: true, schema: {type: string}, description: Cutoff business period; its dates are shifted by zero through three years.}
        - {name: indicatorCode, in: query, required: true, schema: {type: string}, description: Code returned by annual-comparison-definitions; never a browser-owned metric.}
      responses:
        '200':
          description: Four current/prior-year values from approved database records only.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AnnualComparison'
        '403': {description: Requested region is outside the caller read scope.}
components:
  schemas:
    AnnualComparisonDefinition:
      type: object
      required: [code, name, unitCode, sourceDomain, aggregationCode]
      properties:
        code: {type: string}
        name: {type: string}
        unitCode: {type: string}
        sourceDomain: {type: string, enum: [PRODUCTION, MARKET]}
        aggregationCode: {type: string, enum: [SUM, AVERAGE]}
    AnnualComparison:
      type: object
      required: [indicatorCode, indicatorName, sourceDomain, productCode, regionCode, cutoffPeriodCode, unitCode, methodologyVersion, points]
      properties:
        indicatorCode: {type: string}
        indicatorName: {type: string}
        sourceDomain: {type: string, enum: [PRODUCTION, MARKET]}
        productCode: {type: string}
        cultivarCode: {type: string, nullable: true}
        regionCode: {type: string}
        cutoffPeriodCode: {type: string}
        unitCode: {type: string}
        methodologyVersion: {type: string, example: OVERVIEW_APPROVED_FACTS_V2}
        points:
          type: array
          minItems: 4
          maxItems: 4
          items:
            type: object
            required: [businessYear, value, sourcePublicationVersion, dataCutoff, missingReason]
            properties:
              businessYear: {type: string, example: '2026'}
              value: {type: number, nullable: true}
              sourcePublicationVersion: {type: string, nullable: true}
              dataCutoff: {type: string, format: date-time, nullable: true}
              missingReason: {type: string, nullable: true, enum: [NO_APPROVED_RECORDS]}